ct/l1: rename frontend to leader_router
